+++ Realm/Realm/Handles/SharedRealmHandle.cs
@@ -117,6 +117,10 @@ namespace Realms
 
             [DllImport(InteropConfig.DLL_NAME, EntryPoint = "shared_realm_install_callbacks", CallingConvention = CallingConvention.Cdecl)]
             public static extern void install_callbacks(NotifyRealmCallback notifyRealmCallback, GetNativeSchemaCallback nativeSchemaCallback);
+
+            [DllImport(InteropConfig.DLL_NAME, EntryPoint = "shared_realm_has_changed", CallingConvention = CallingConvention.Cdecl)]
+            [return: MarshalAs(UnmanagedType.I1)]
+            public static extern bool has_changed(SharedRealmHandle sharedRealm);
         }
 
         static SharedRealmHandle()
@@ -298,6 +302,11 @@ namespace Realms
                 default:
                     throw new NotSupportedException($"Unexpected primary key of type: {pkProperty.Type}");
             }
+        }
+
+        public bool HasChanged()
+        {
+            return NativeMethods.has_changed(this);
         }
 
         private ObjectHandle CreateObjectWithPrimaryKey(TableHandle table, long? key, bool isNullable, bool update, out bool isNew)